refactor : Sort component
Related to : #42
