+++ src/app/best/page.tsx
@@ -1,15 +1,10 @@
 import Category from '@/components/best/category/Category';
-import BookItem from '@/components/common/BookItem';
+import Rank from '@/components/best/rank/Rank';
 import PageHeader from '@/components/common/PageHeader';
 import Sort from '@/components/common/Sort';
 import styles from '@/styles/best/best.module.css';
 
 export default function bestPage() {
-	let ranks = [];
-	for (let i = 1; i <= 30; i++) {
-		ranks.push(i);
-	}
-
 	return (
 		<>
 			<PageHeader title="베스트셀러" />
@@ -17,14 +12,7 @@ export default function bestPage() {
 				<div className={styles.wrapper}>
 					<Category />
 					<Sort />
-					<div className={styles.rankBox}>
-						{ranks.map((rank) => (
-							<div className={styles.itemWrapper} key={rank}>
-								<div className={styles.rank}>{rank}</div>
-								<BookItem />
-							</div>
-						))}
-					</div>
+					<Rank />
 				</div>
 			</div>
 		</>
