+++ src/styles/best/best.module.css
@@ -18,25 +18,3 @@
 .wrapper > :last-child {
 	margin-bottom: 100px;
 }
-
-/* Rank */
-.rankBox {
-	width: inherit;
-	column-gap: 20px;
-
-	display: grid;
-	grid-template-columns: repeat(6, 1fr);
-}
-.itemWrapper {
-	height: 500px;
-}
-.rank {
-	width: inherit;
-	height: 90px;
-	font-size: 26px;
-	color: #6594e7;
-
-	display: flex;
-	align-items: center;
-	justify-content: center;
-}
