+++ src/styles/best/rank/rank.module.css
@@ -1,0 +1,21 @@
+/* Rank */
+.rankBox {
+	width: inherit;
+	column-gap: 20px;
+
+	display: grid;
+	grid-template-columns: repeat(6, 1fr);
+}
+.itemWrapper {
+	height: 500px;
+}
+.rank {
+	width: inherit;
+	height: 90px;
+	font-size: 26px;
+	color: #6594e7;
+
+	display: flex;
+	align-items: center;
+	justify-content: center;
+}
